requirements.txt
fastapi[standard]
uvicorn
pymongo
python-dotenv
requests
python-multipart